requirements.txt
flask
pycryptodome
requests
protobuf